perf(lift): make the eager replay cheap — 28k → 8.3k instructions per hit (#40)
* perf(core): pack the ids and links an instruction carries

An `Instruction` was 160 bytes, a `Mnemonic` 80 and an operand 16, and a
session's body on /usr/bin/bash holds two million instructions. Now 112,
56 and 8:

- The module-interned ids (`LiteralId`, `VarnodeId`, `BytesId`,
  `PoisonId`) and the body-local temporary ids are `u32`, as the other
  body-local ids already were; a `LocalValueId` is eight bytes.
- The links an instruction carries — its block, its neighbours, its
  first use — and its machine address are packed (`value::link`): `None`
  is the one pattern no arena issues. Reads and writes go through
  `Option`, serde sees the same `Option`, and `WithUsers` sets a head
  instead of lending it.
- The argument lists of terminators, calls, intrinsics, p-code ops and
  tuples are `Box<[LocalValueId]>` rather than `Vec`: a mnemonic never
  grows one in place, and the box is a third smaller.

Peak memory of a session over bash: 695 → 560 MB. No change to the IR.

* perf(sleigh): resolve a template's types and fixed constants once per session

Most of a template's constants are the same at every instance — sizes,
flag values — and every replay interned each of them again, a hash probe
under the module's lock per constant. The session now keeps, per
template it has replayed, what the template's types and fixed constants
resolve to in its context; a replay interns only the constants that
move with the address or a parameter.

* perf(core): keep function-local names as a base and a suffix

Nearly every name a lifted body holds is derived: a load named after its
register (`rax`, `rax_1`, `rax_2`, …) or a block labelled with its
address. Each was a heap string, minted through a string-keyed table.

A function-local `Name` is now eight bytes — a `BaseId` into the body's
interned bases plus a suffix — rendered as `base` or `base_<n>` on the
way out; a block at an address with no name of its own holds none and
renders as the address. `value::name::LocalNames` keeps the table as a
vector per base indexed by suffix (with a far map for explicit outliers)
and the address labels by value, so minting, registering and resolving
never hash a suffixed string, and growing the table never re-reads one.
The uniqueness rule is the flat-table one: first free `base_<n>` with
`n ≥ 1`, in creation order; an explicit `base_<n>` takes its slot; a
label and an explicit name spelled alike exclude each other.

`Named::name` returns `Cow<str>`. `BlockParam::name` is crate-private,
set through the body; `Temp::with_name` becomes
`FunctionBody::push_named_temp`. On the wire a body carries its bases and
each value its `Name`; the table is rebuilt on load like the use edges.
`Builder::push_mnemonic_with_type_based` takes an interned base, which
the cache replay memoizes per template and session.

Instruction 112 → 96 bytes. Replaying a hit on /usr/bin/bash: 14.4k →
11.0k retired instructions; peak memory 558 → 484 MB.

* perf(lift): append a replayed instruction as one prototype run

A cache hit was emitted one operation at a time through the builder, each
paying the builder's bookkeeping and rebuilding what the template already
knew. `FunctionBody::append_prototype` takes the whole run — the
template's operations with the session's type ids and name bases, made
once per template and session (`Resolved`) — and does per operation only
what the body must: copy the mnemonic, point its operands into the body,
push, link, record the use edges, connect the branches, name the result.

That also fixes a replay bug: a replayed branch made no CFG edge, so the
blocks of a session's hits had no successors. A new test checks the
successor, predecessor and use counts of a session of hits against a
session of misses.

Along the way: the use-list heads of literals and varnodes are dense
vectors instead of a hash map (`SharedHeads`); `add_use` matches once
without dynamic dispatch; the operand walks are inlined; a commit no
longer scans an instruction's operations for minted callees when it
promised none and reports no call.

Replaying a hit on /usr/bin/bash: 11.0k → 9.4k retired instructions,
7.4k → 6.5k cycles.

* perf(core): keep a block's incident edges inline

A block's edge set was a `HashSet` — an allocation and a hasher per
block, and a flat lift makes a block per machine instruction. A block has
one or two successors and a few predecessors, so `EdgeSet` holds the ids
inline (three before spilling) and looks one up by scanning. Iteration is
in insertion order, which is deterministic, as the fixed-seed hasher was.

Replaying a hit on /usr/bin/bash: 9.4k → 8.8k retired instructions.

* refactor(sleigh): pass the lifter, decoder and flatness as one Lowering

The cache's lookup, lower and miss each took the three the caller lowers
with — the lifter, the decoder that decoded the instruction and decodes
its probes, and whether control flow is flattened — plus a scratch. One
`Lowering` carries them, and the session builds it once per lift.

* perf(core): cheaper block labels and one list update per run

Two per-instruction costs of a flat lift, which makes a block and a
handful of operations for every machine instruction:

- Labelling a block with its address built the address's spelling just to
  ask whether a name spells it too. Each interned base now knows the
  address it spells, and the bases that spell one are listed by it, so a
  label is one lookup and no allocation.
- Appending a recorded run read and wrote a block's instruction list per
  operation. The operations are grouped by block, so the list is now read
  once per run of one block and written once, with the operations linked
  to each other directly; an instruction arrives fully linked
  (`Instruction::linked`) rather than built and then patched.

Replaying a hit on /usr/bin/bash: 8.8k → 8.3k retired instructions.

* docs(sleigh): record what made the eager replay cheap, and what is left

* docs(core): do not link a private verb from the builder's

* docs: hand over the replay speed work

Where the branch stands, how to measure it on a machine that is never
idle, where the remaining 583 instructions per operation and 3 300 fixed
go, and what has been measured and rejected — a reserve on jstd's arenas
among them, which costs slightly more than it saves.

+++ core/src/value/poison.rs
@@ -28,7 +28,7 @@ use crate::{
 use jstd::Identifier;
 
 #[derive(Identifier)]
-pub struct PoisonId(usize);
+pub struct PoisonId(u32);
 
 /// A typed poison value stored in a [`Context`](crate::context::Context). Carries
 /// only its [`TypeId`] (hence its width); its bits are undefined.
